CQC template: Changed handling of sides and counts...
It is now based on sides, rather than groups or a variable.
Players begin as civilian, and get side-switched when they teleport out.
They are switch back to civilian when they RTB.
The count simply counts all players in the given side.

This commit also fixes an error from the previous one.